Refactored AccuracyEstimator into three objects
The old way I was doing it was rapidly becoming untenable. Now it uses abstraction and inheritance, like a good program ought to.
