+++ src/main/java/drgtools/dpscalc/weapons/gunner/minigun/Minigun_SpreadCurve.java
@@ -1,7 +1,7 @@
 package drgtools.dpscalc.weapons.gunner.minigun;
 
 import drgtools.dpscalc.guiPieces.accuracyEstimator.LineGraph;
-import drgtools.dpscalc.weapons.SpreadCurve;
+import drgtools.dpscalc.modelPieces.accuracy.SpreadCurve;
 
 public class Minigun_SpreadCurve extends SpreadCurve {
 	@Override
