+++ src/main/java/drgtools/dpscalc/weapons/scout/classic/Classic_SpreadCurve.java
@@ -1,7 +1,7 @@
 package drgtools.dpscalc.weapons.scout.classic;
 
 import drgtools.dpscalc.guiPieces.accuracyEstimator.LineGraph;
-import drgtools.dpscalc.weapons.SpreadCurve;
+import drgtools.dpscalc.modelPieces.accuracy.SpreadCurve;
 
 public class Classic_SpreadCurve extends SpreadCurve {
 	@Override
